requirements.txt
# Core dependencies
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Additional dependencies
alembic==1.13.1
python-dotenv==1.0.0
jinja2==3.1.2
cryptography==41.0.8
stripe==8.6.0
prometheus-client==0.19.0
sentry-sdk==1.38.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0